+++ protocol/judge/commands/__init__.py
@@ -1,24 +1,7 @@
 """
-This module contains the Commands class.
+This module contains the judge commands.
 """
 
-from enum import Enum
+from .judge_commands import Commands
 
-from .check_command import CheckCommand
-from .start_command import StartCommand
-
-
-class Commands(Enum):
-    """
-    This class contains all the commands that the runner can execute.
-    """
-
-    START = StartCommand()
-    """
-    Start a container on the runner.
-    """
-
-    CHECK = CheckCommand()
-    """
-    Check the status of the runner.
-    """
+__all__ = ["Commands"]
