+++ protocol/judge/commands/check_command.py
@@ -1,10 +1,17 @@
+"""
+This module contains the CheckCommand class.
+"""
+
 from .command import Command
 
 
 class CheckCommand(Command):
+    """
+    The CheckCommand class is used to check the status of the runner.
+    """
+
     @staticmethod
     def response(response: dict):
-
         if response["status"] is None:
             raise ValueError("Received message with missing status!")
