+++ protocol/judge/commands/start_command.py
@@ -1,8 +1,15 @@
+"""
+This module contains the StartCommand class.
+"""
+
 from .command import Command
 
 
 class StartCommand(Command):
+    """
+    The StartCommand class is used to start a container on the runner.
+    """
+
     @staticmethod
     def response(response: dict):
         pass
-
